refactor: Move `linux_checks_librs` to own module
Part of #329
