+++ src-tauri/src/lib.rs
@@ -9,9 +9,6 @@ mod platform_specific;
 #[cfg(target_os = "windows")]
 use platform_specific::windows;
 
-#[cfg(target_os = "linux")]
-use platform_specific::linux;
-
 use serde::{Deserialize, Serialize};
 use sysinfo::SystemExt;
 use tokio::time::sleep;
@@ -82,28 +79,6 @@ pub fn check_mod_version_number(path_to_mod_folder: &str) -> Result<String, anyh
     log::info!("{}", mod_version_number);
 
     Ok(mod_version_number.to_string())
-}
-
-// I intend to add more linux related stuff to check here, so making a func
-// for now tho it only checks `ldd --version`
-// - salmon
-#[cfg(target_os = "linux")]
-pub fn linux_checks_librs() -> Result<(), String> {
-    // Perform various checks in terms of Linux compatibility
-    // Return early with error message if a check fails
-
-    // check `ldd --version` to see if glibc is up to date for northstar proton
-    let min_required_ldd_version = 2.33;
-    let lddv = linux::check_glibc_v();
-    if lddv < min_required_ldd_version {
-        return Err(format!(
-            "GLIBC is not version {} or greater",
-            min_required_ldd_version
-        ));
-    };
-
-    // All checks passed
-    Ok(())
 }
 
 /// Attempts to find the game install location
